# Limit ranked jobs