--- a/scheduler/test/cook/test/benchmark.clj
+++ b/scheduler/test/cook/test/benchmark.clj
@@ -16,7 +16,8 @@
 
 (ns cook.test.benchmark
   (:use clojure.test)
-  (:require [cook.scheduler.dru :as dru]
+  (:require [cook.quota :as quota]
+            [cook.scheduler.dru :as dru]
             [cook.scheduler.scheduler :as sched]
             [cook.scheduler.share :as share]
             [cook.tools :as util]
@@ -54,10 +55,12 @@
         (cc/quick-bench (sched/rank-jobs db offensive-job-filter))))
     (testing "sort-jobs-by-dru-helper"
       (let [db (d/db conn)
-            pending-task-ents (util/get-pending-job-ents db)
+            pending-task-ents (->> (util/get-pending-job-ents db)
+                                   (map util/create-task-ent))
             running-task-ents (util/get-running-task-ents db)
             sort-task-scored-task-pairs dru/sorted-task-scored-task-pairs
-            user->dru-divisors (share/create-user->share-fn db nil)]
+            user->dru-divisors (share/create-user->share-fn db nil)
+            user->quota (quota/create-user->quota-fn db nil)]
         (do
           (println "============ sort-jobs-by-dru timing ============")
           (cc/quick-bench (sched/sort-jobs-by-dru-helper pending-task-ents
@@ -65,5 +68,6 @@
                                                          user->dru-divisors
                                                          sort-task-scored-task-pairs
                                                          (timers/timer (sched/metric-title "sort-jobs-hierarchy-duration" "no-pool"))
-                                                         "no-pool"))
+                                                         "no-pool"
+                                                         user->quota))
           nil)))))