## Changes proposed in this PR
- Limit the number of over-quota jobs output by ranking

## Why are we making these changes?
This can cause performance issues downstream in a situation where there are relatively few runnable jobs, but many jobs that are over quota in the global queue.

After change:
```
lein test cook.test.benchmark
============ rank-jobs timing ============
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 842.769115 ms
    Execution time std-deviation : 32.152188 ms
   Execution time lower quantile : 819.166167 ms ( 2.5%)
   Execution time upper quantile : 892.687703 ms (97.5%)
                   Overhead used : 1.849735 ns

Found 1 outliers in 6 samples (16.6667 %)
	low-severe	 1 (16.6667 %)
 Variance from outliers : 13.8889 % Variance is moderately inflated by outliers
============ rank-jobs minus offensive-job-filter timing ============
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 691.942744 ms
    Execution time std-deviation : 5.447180 ms
   Execution time lower quantile : 683.694154 ms ( 2.5%)
   Execution time upper quantile : 697.856670 ms (97.5%)
                   Overhead used : 1.849735 ns
============ sort-jobs-by-dru timing ============
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 282.943425 ms
    Execution time std-deviation : 5.670062 ms
   Execution time lower quantile : 275.517746 ms ( 2.5%)
   Execution time upper quantile : 289.528243 ms (97.5%)
                   Overhead used : 1.849735 ns
```

Before change:

```
============ rank-jobs timing ============
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 760.793033 ms
    Execution time std-deviation : 22.661330 ms
   Execution time lower quantile : 743.616719 ms ( 2.5%)
   Execution time upper quantile : 793.321066 ms (97.5%)
                   Overhead used : 1.519518 ns
============ rank-jobs minus offensive-job-filter timing ============
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 680.558862 ms
    Execution time std-deviation : 5.242206 ms
   Execution time lower quantile : 670.552910 ms ( 2.5%)
   Execution time upper quantile : 684.407241 ms (97.5%)
                   Overhead used : 1.519518 ns

Found 1 outliers in 6 samples (16.6667 %)
	low-severe	 1 (16.6667 %)
 Variance from outliers : 13.8889 % Variance is moderately inflated by outliers
============ sort-jobs-by-dru timing ============
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 370.130333 ms
    Execution time std-deviation : 13.754199 ms
   Execution time lower quantile : 353.049737 ms ( 2.5%)
   Execution time upper quantile : 384.385778 ms (97.5%)
                   Overhead used : 1.519518 ns
```